requirements.txt
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
dnspython==2.2.1
firebase-admin==5.2.0
google-api-core==2.7.1
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
lxml==4.8.0
msgpack==1.0.3
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
pika==1.2.1
pika-stubs==0.1.3
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pymongo==3.12.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
types-PyMySQL==1.0.13
types-pytz==2021.3.5
types-redis==4.1.17
types-requests==2.27.11
types-urllib3==1.26.10
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.14.0
websocket-client==1.4.0
humanfriendly